feat: try using non-zero realm/shard (#18092)

+++ hedera-node/test-clients/src/main/resources/contract/contracts/JustSend/JustSend.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.0;
 
 contract JustSend {
-    function sendTo(uint64 account_num, uint64 value) public payable {
+    function sendTo(uint160 account_num, uint64 value) public payable {
         address payable beneficiary = payable(address(uint160(account_num)));
         beneficiary.transfer(value);
     }
