+++ hedera-node/test-clients/src/main/resources/validation-scenarios/Multipurpose.sol
@@ -17,7 +17,7 @@ contract Multipurpose {
     return luckyNumber;
   }
 
-  function donate(uint32 toNum, string memory saying) public payable {
+  function donate(uint120 toNum, string memory saying) public payable {
     address payable beneficiary = address(uint120(toNum));
     beneficiary.transfer(1);
     emit Boast(saying);
